+++ tests/org.preesm.tests.integration/resources/org.ietr.preesm.loopgen-sobel-erosion-dilation/CodeMPPA2/include/preesm.h
@@ -14,63 +14,40 @@
 
 #include <stdlib.h>
 #include <stdio.h>
-#include <string.h>
 
 // Using nodeos POSIX pthread are available 
 #include <pthread.h>
-#include <semaphore.h>
 
-#ifndef __k1io__
-#include "communication.h"
-#include "fifo.h"
-#include "clock.h"
-#include "splitMerge.h"
+#if 1
+#define GENERATE_FILE
+#define PREESM_LOOP_SIZE 300
 #endif
 
-#define SHARED_SEGMENT_ID (10)
-
-#ifdef __bare__
-#include "utask.h"
-#else
+#if 0
+//d#define GENERATE_FILE
+#define PREESM_LOOP_SIZE 3
 #endif
-#define NB_OMP_CORE (16)
 
-#define GRAPH_ITERATION 200
-
-#define VERBOSE
+#define _PREESM_PAPIFY_MONITOR
 
 //#define DISABLE_IO
-
-//#define PATH "/home/user/Temp/akiyo_cif.yuv"
-//#define PATH "/home/accesscore/jhascoet/akiyo_cif.yuv"
-#define PATH "/home/accesscore/jhascoet/akiyo_cif_640x480.yuv"
-#if 0
-#define VIDEO_WIDTH 352
-#define VIDEO_HEIGHT 288
-#endif
-#if 1
-#define VIDEO_WIDTH 640
-#define VIDEO_HEIGHT 480
-#endif
 
 #define IN
 #define OUT
 
 //#define PROFILE
-#define NB_CORE 1
+#define PREESM_NB_CORES_CC 10
+#define PREESM_NB_CORES_IO 3
 
 /* user include */
-#ifndef __k1io__
 #include "sobel.h"
 #include "morph.h"
 #include "medianFilter.h"
 #include "yuvRead.h"
 #include "yuvDisplay.h"
-#endif
 
 /* user defined definition */
 //#define Shared ((char*)0x90000000ULL) 	/* Shared buffer in DDR */
-#define Shared (0x90000000ULL) 	/* Shared buffer in DDR */
 
 typedef unsigned char uchar;