+++ release_notes.md
@@ -5,15 +5,17 @@ PREESM Changelog
 *XXXX.XX.XX*
 
 ### New Feature
+*  Clustering and DistributedOnly memory distribution now are compatible;
 
 ### Changes
 *  Fix C Hardware codegen: properly prefix LOOP_SIZE with PREESM_;
 * SpiderCodegen: updated spider codegeneration to support the new Archi model of Spider.
 *  Fix C Hardware codegen: properly printed the Release Kernel Instance with the name of the function to be executed in Hardware
 *  Update workflow editor to better report task id issues;
+*  Changed _PREESM_MONITOR_INIT by _PREESM_PAPIFY_MONITOR;
 
 ### Bug fix
-
+*  Fix several bugs on the communications when using the IO in the MPPA code generation;
 
 ## Release version 3.7.0
 *2019.05.14*